+++ src/main/scala/dxWDL/Main.scala
@@ -54,8 +54,7 @@ object Main extends App {
                                billTo: String,
                                compileMode: CompilerFlag.Value,
                                defaults: Option[Path],
-                               dxWDLrtId: String,
-                               lockedWf: Boolean,
+                               locked: Boolean,
                                inputs: List[Path],
                                region: String,
                                reorg: Boolean,
@@ -424,7 +423,6 @@ object Main extends App {
     private def compilerOptions(options: OptionsMap, bOpt: BaseOptions) : CompilerOptions = {
         // get billTo and region from the project
         val (billTo, region) = Utils.projectDescribeExtraInfo(bOpt.dxProject)
-        val dxWDLrtId = getAssetId(region)
         val compileMode: CompilerFlag.Value = options.get("compilemode") match {
             case None => CompilerFlag.Default
             case Some(List(x)) if (x.toLowerCase == "ir") => CompilerFlag.IR
@@ -449,7 +447,6 @@ object Main extends App {
                         billTo,
                         compileMode,
                         defaults,
-                        dxWDLrtId,
                         options contains "locked",
                         inputs,
                         region,
@@ -519,7 +516,8 @@ object Main extends App {
         // unmodified.
         val cef = new CompilerErrorFormatter(ns.terminalMap)
         var irNs = CompilerIR(bOpt.folder, instanceTypeDB, cef,
-                              cOpt.reorg, cOpt.lockedWf, bOpt.verbose).apply(ns)
+                              cOpt.reorg, cOpt.locked, bOpt.verbose).apply(ns)
+        val allStageNames = irNs.workflow.get.stages.map{ stg => stg.name }.toVector
 
         irNs = cOpt.defaults match {
             case Some(path) =>
@@ -527,6 +525,10 @@ object Main extends App {
                 InputFile(bOpt.verbose).embedDefaults(irNs, path)
             case _ => irNs
         }
+
+        // make sure the stage order hasn't changed
+        val embedAllStageNames = irNs.workflow.get.stages.map{ stg => stg.name }.toVector
+        assert(allStageNames == embedAllStageNames)
 
         // Write out the intermediate representation
         prettyPrintIR(wdlSourceFile, irNs, bOpt.verbose.on)
@@ -552,10 +554,11 @@ object Main extends App {
         val wf:Option[DXWorkflow] = cOpt.compileMode match {
             case CompilerFlag.Default =>
                 // Generate dx:applets and dx:workflow from the IR
+                val dxWDLrtId = getAssetId(cOpt.region)
                 val (wf, _) =
-                    CompilerNative(cOpt.dxWDLrtId, bOpt.dxProject, instanceTypeDB,
+                    CompilerNative(dxWDLrtId, bOpt.dxProject, instanceTypeDB,
                                    bOpt.folder, cef,
-                                   bOpt.force, cOpt.archive, cOpt.lockedWf, bOpt.verbose).apply(irNs)
+                                   bOpt.force, cOpt.archive, cOpt.locked, bOpt.verbose).apply(irNs)
                 wf
             case CompilerFlag.IR =>
                 None